feat: Added overrides for invite check
- Overrides for Moderator+

+++ bot/cogs/utility/links.py
@@ -63,10 +63,13 @@ class Links(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message: discord.Message):
+        if any(role_check in ['Moderator','Administrator'] for role_check in [role.name for role in message_object.author.roles]):
+            return
+        
         if self.bot.user == message.author:
             return
         words = ["discord.gg/", "com/invite/"]
-
+        
         for word in words:
             if word in message.content:
                 return await message.delete()
